Add calendario

+++ src/AppBundle/Controller/SedeController.php
@@ -6,6 +6,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Component\HttpFoundation\Request;
 use AppBundle\Entity\Sede;
+use AppBundle\Entity\Permisos;
 use AppBundle\Form\SedeType;
 
 /**
@@ -38,6 +39,17 @@ class SedeController extends Controller
         if ($form->isSubmitted() && $form->isValid()) {
             $em->persist($sede);
             $em->flush();
+            $permisosUser = new Permisos;
+            $permisosUser->setSede($sede);
+            $permisosUser->setType('USER');
+            $permisosUser->setPermisos([]);
+            $permisosAdmin = new Permisos;
+            $permisosAdmin->setSede($sede);
+            $permisosAdmin->setType('ADMIN');
+            $permisosAdmin->setPermisos([]);
+            $em->persist($permisosUser);
+            $em->persist($permisosAdmin);
+            $em->flush();
             return $this->redirectToRoute('sedes_index');
         }
         return $this->render('AppBundle:Sede:new.html.twig', array(
@@ -69,6 +81,18 @@ class SedeController extends Controller
     public function delAction(Sede $sede)
     {
         $em = $this->getDoctrine()->getManager();
+        $user = $em->getRepository('AppBundle:Users')
+        ->findBySede($sede->getId()); 
+        $child = $em->getRepository('AppBundle:Childs')
+        ->findBySede($sede->getId()); 
+        if ($user) {
+            $this->addFlash('notice','Hy usuarios registrados con esta sede');
+            return $this->redirectToRoute('sedes_index');
+        }
+        if ($child) {
+            $this->addFlash('notice','Hay registros con esta sede ');
+            return $this->redirectToRoute('sedes_index');
+        }
         $em->remove($sede);
         $em->flush();
         return $this->redirectToRoute('sedes_index');
@@ -84,7 +108,7 @@ class SedeController extends Controller
         ->findOneBy(['sede'=>$sede->getId(),'type'=>'USER']); 
         $permisosAdmin = $em->getRepository('AppBundle:Permisos')
         ->findOneBy(['sede'=>$sede->getId(),'type'=>'ADMIN']); 
-        if ($request->get('permisosUser')) {
+        if ($request->get('permisosUser') || $request->get('permisosAdmin') ) {
             $permisosUser->setPermisos($request->get('permisosUser'));
             $permisosAdmin->setPermisos($request->get('permisosAdmin'));
         $em->flush();
